packages/ghappkit-client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ghappkit_client"]

[project]
name = "ghappkit-client"
version = "0.1.0"
description = "GitHub auth/client integration package for ghappkit."
requires-python = ">=3.10.0"
dependencies = [
  "httpx>=0.27",
  "PyJWT[crypto]>=2.9",
  "cryptography>=42",
]
